secproto/requirements.txt
fastapi
uvicorn[standard]
python-multipart
cryptography
SQLAlchemy
pydantic
python-dotenv
PyJWT
passlib[bcrypt]
